resolves issue #3 (attempt)
- Remove old temp_layer ids stored in the forms.
- Soft code look-up relations for different classes.

DISCLAMER: look-up table approach is not yet refined.
Currently the apporach changed from importing 'lu_%' tables
to getting the enumeration and codelist tables.

For the enumeration, relative to terrain and water ids are hardcoded,
because there isn't any key present in layer_metadata or elsewhere.

For the codelist, the code_ids are filtered by a resemblence between
the name of the codelist, the type (Class,Function,Usage) and the root_class of the layer.
That said, this apporach doesn't work for '%Part' and additional types like 'species'