Temporary fix of the "false" message fire that extents don't ovelap.
Why it happens:
1.The basemap is initialized with bbox 0,0,0,0
2.This sends a signal that the extents of the basemap have been changed.
3.When the extents are changed, the code checks if these new extents
overlap with the extents of the layers coming from the database.
So in this case these two extents don't ovelap causing the message to
appear.

4.After that the code assign the layer extents to the basemap.
5.And the process repeats from step 2
6.Only this time, both extents are the same -one on top of the other-
which is what we want.

How it is currently fixed:
In the conditional statement I added that in the case of basemap extent
of |0,0,0,0|, do nothing, as it means that these are the initialization
extents which don't count as a proper extent change.

This kind of fix is temporar.